internal: refactor NoSuchField diagnostic
